refactor(ui): extract StatTile primitive; delete dead MetricCard
Establish src/components/ui/StatTile.tsx as the shared sidebar stat-cell
primitive (label-over-value, interactive view-switch / readOnly readout
variants) by relocating SidebarViewSwitch's local StatCell verbatim — same
class strings, so the rendered output is byte-identical. This is the base
layer for a presentational tile + per-context value hooks: the screen
models decide *what* each number is, StatTile only renders it.

Delete src/components/ui/MetricCard.tsx (+ its MetricGrid). It had zero
render sites — the live tiles use the orange top-bar / inset-bar skins,
not its glass capsule — so it was dead code. The glass-capsule reference
it documented is already covered by the live SelectableCard / FilterCard /
Toolbar primitives; retarget the ~20 prose/doc references accordingly
(no runtime change).

Validation: refactor only — tsc introduces no new errors; pnpm test (122
files) passes. Output is class-identical to before, so no display change
and no version bump.
