+++ src/components/ui/StatTile.tsx
@@ -1,0 +1,83 @@
+import type { ReactNode } from "react";
+
+// Presentational stat tile for the sidebar's joined "hero stats" blocks — the
+// label-over-value cell that fills the bordered footer rows under the flight
+// count (airport + here mode) and, going forward, the flight-tracking
+// telemetry grid. It is intentionally dumb: callers pass an already-formatted
+// value (NumberFlow, an em dash, plain text) plus the label/unit/prefix; the
+// per-context "what is this number" logic lives in the screen models/hooks, not
+// here.
+//
+// Two modes share one skin:
+//   - interactive (default, has `onClick`): a tab-like button with the orange
+//     top-rail on `active` and a hover wash — used for view switches and the
+//     here-mode km/h⇄mph speed toggle.
+//   - readOnly: a plain cell with no hover/active affordance, so pure readouts
+//     (here-mode altitude) don't read as tappable.
+//
+// This supersedes the old ui/MetricCard tile primitive (which carried the glass
+// capsule but was never rendered); the live glass-capsule reference now lives in
+// SelectableCard / FilterCard / Toolbar.
+type StatTileProps = {
+  label: ReactNode;
+  value: ReactNode;
+  unit?: ReactNode;
+  prefix?: ReactNode;
+  active?: boolean;
+  onClick?: () => void;
+  readOnly?: boolean;
+};
+
+export default function StatTile({
+  label,
+  value,
+  unit,
+  prefix,
+  active,
+  onClick,
+  readOnly = false,
+}: StatTileProps) {
+  const body = (
+    <>
+      <div className="truncate text-[calc(10px*var(--sb-body-scale))] text-atc-faint">{label}</div>
+      <div className="mt-[3px]">
+        {prefix ? (
+          <span
+            className="notranslate text-[calc(10px*var(--sb-body-scale))] text-atc-faint"
+            translate="no"
+          >
+            {prefix}
+          </span>
+        ) : null}
+        <span className="text-[calc(16px*var(--sb-body-scale))] font-normal tabular-nums text-atc-text">
+          {value}
+        </span>
+        {unit ? (
+          <span className="ml-0.5 text-[calc(10px*var(--sb-body-scale))] text-atc-faint">{unit}</span>
+        ) : null}
+      </div>
+    </>
+  );
+
+  // Read-only stats (here-mode altitude) are not view switches: render a plain
+  // cell with no hover/active affordance so they don't read as tappable.
+  if (readOnly) {
+    return (
+      <div className="relative min-w-0 flex-1 px-[11px] py-[9px] text-left [&:not(:last-child)]:border-r [&:not(:last-child)]:border-[var(--app-frost-border)]">
+        {body}
+      </div>
+    );
+  }
+
+  return (
+    <button
+      type="button"
+      data-active={active ? "true" : undefined}
+      onClick={onClick}
+      aria-pressed={active}
+      className="relative min-w-0 flex-1 px-[11px] py-[9px] text-left transition-[background-color] duration-200 ease-out [&:not(:last-child)]:border-r [&:not(:last-child)]:border-[var(--app-frost-border)] before:pointer-events-none before:absolute before:inset-x-0 before:top-0 before:h-[2px] before:origin-center before:scale-x-0 before:bg-[var(--atc-signal-accent)] before:transition-transform before:duration-300 before:ease-[cubic-bezier(0.34,1.3,0.64,1)] hover:bg-[var(--atc-control-hover-bg)] data-[active=true]:bg-[color-mix(in_oklab,var(--atc-signal-accent)_11%,transparent)] data-[active=true]:before:scale-x-100"
+    >
+      {body}
+    </button>
+  );
+}
